Start spiking agnostic way to ssr web components with shadow roots.
