+++ packages/sk-shadow/src/register.ts
@@ -1,0 +1,3 @@
+import Shadow from '.';
+
+customElements.define('sk-shadow', Shadow);
